Initial gif search
This initial commit adds the most basic functionality:
- Show the trending gifs from giphy

This adds a basic react app with an app that uses an api_key to
fetch the trending gifs as json. This is used to display the images
in an Image react component.

This is a proof-of-concept of fetching the data and will be expanded
upon.