Add basic search functionality
This commit adds a basic search input that will submit the query to
giphy and show the results.

This adds new styles to put the search bar at the top of the page
in the titlebar area.

This modifies the App to abstract the adding of images to a new function
setImages(), and uses this in a new function searchImages(), which acts
like getImages() but takes in a query.

This adds a search component which shows a search bar and calls a search
function passed down from the app.