Break the components out into various files
The main App.js file was getting uncomfortably large. This commit fixes
that by breaking the components out into more structured files.

The buttons used for navigation are in Buttons.js. The search box component
is in Search.js, and the various ways to show images are in Image.js.